Updated .. Model 269 now contains values for damping and pushing that look promising.
In main.cpp I added a pushing feature, because the rotating point is not a perfect point, it has friction which causes the second part of the pendulum to get pushed a little to the right..
Just a very little bit, but by adding it you get the characteristic shape of this figure.
And: It's now possible to trim it to the research setup.